ip_to_pkg: Move library between diretories
