+++ modules/dnsmasq_dhcp/dhcp.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/netdata/go-orchestrator/module"
 
-	"github.com/netdata/go.d.plugin/modules/dnsmasq_dhcp/ip"
+	"github.com/netdata/go.d.plugin/pkg/ip"
 )
 
 func init() {
